requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
idna==3.4
jwt==1.3.1
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tb-rest-client==3.6
urllib3==2.0.7
